feat: loading state for filters
