+++ src/app/(light)/explorer/data-models/components/search/fields/amount-of-issuances-field.tsx
@@ -1,18 +1,19 @@
 'use client';
 
-import RangeField from '@/components/form/range-field/range-field';
+import RangeField, {
+  RangeFieldProps,
+} from '@/components/form/range-field/range-field';
 import { explorerDataModels } from '@/locale/en/datamodel';
 
 type Props = {
-  min?: number;
-  max: number;
   selectedAmountOfIssuances: number[];
   setAmountOfIssuances: (amountOfIssuances: number[]) => void;
-};
+} & Pick<RangeFieldProps, 'min' | 'max' | 'isLoading'>;
 
 export default function AmountOfIssuancesField({
   min = 0,
   max,
+  isLoading,
   selectedAmountOfIssuances,
   setAmountOfIssuances,
 }: Props) {
@@ -26,6 +27,7 @@ export default function AmountOfIssuancesField({
       value={selectedAmountOfIssuances}
       min={min}
       max={max}
+      isLoading={isLoading}
     />
   );
 }
