--- src/app/(light)/explorer/data-models/components/search/use-metadata.ts
@@ -1,12 +1,0 @@
-'use client';
-
-import { apiPublic } from '@/services/protocol/api';
-import { useQuery } from '@tanstack/react-query';
-
-export default function useMetadata() {
-  const metadata = useQuery({
-    queryKey: ['data-models-metadata'],
-    queryFn: () => apiPublic.explorer_data_models_metadata(),
-  });
-  return metadata;
-}
